+++ tests/Browser/Visits/PagesVisitsAsLenderTest.php
@@ -85,4 +85,14 @@ class PagesVisitsAsLenderTest extends DuskTestCase
                 ->assertSee(403);
         });
     }
+
+    public function testAccountPageVisit()
+    {
+        $this->browse(function (Browser $browser) {
+            $browser->loginAs($this->lender)
+                ->visit('/account')
+                ->assertPathIs('/account')
+                ->assertSee(__('messages.titles.account'));
+        });
+    }
 }